Moved tests from web project to web.tests and ported to xunit
