--- Source/VolunteerReporting/Tests/Web.Tests/TestClass.cs
@@ -1,25 +1,0 @@
-using System;
-using Xunit;
-
-namespace Web.Tests
-{
-    public class TestClass
-    {
-        [Fact]
-        public void PassingTest()
-        {
-            Assert.Equal(4, Add(2, 2));
-        }
-
-        [Fact]
-        public void FailingTest()
-        {
-            Assert.Equal(5, Add(2, 2));
-        }
-
-        int Add(int x, int y)
-        {
-            return x + y;
-        }
-    }
-}
